requirements.txt
numpy
pandas
matplotlib
pyserial
opencv-python
opencv-python-headless
fastapi
pyyaml
